PO-3 - add DTOs for such DB entities: user, wish_list, product, product_price, store. (#10)
* PO-3 - added UserDto and UserRoleEnum classes.

* PO-3 - developing DTOs.

* PO-3 - Finished UserDto and all it staff.

* PO-3 - added additional interface in DtoConverter hierarchy.

* PO-3 - fixed a mistake.

* PO-3 - added unit test for UserDtoConverterImpl, changed user_role id type to TINYINT and reordered UserRoleEnum according to order in DB.

* PO-3 - added StoreDto.

* PO-3 - started development of ProductDto.

* PO-3 - added equals() and hashcode to User and UserRole.

* PO-3 - decided to use Spring library to copy properties from DTO and Entity and vise versa.

* PO-3 - added static import.

* PO-3 - updated hashcode() methods to be ready to null fields, added serializable to DTO's.

* PO-3 - added ProductPropertiesDto and logic for it.

* PO-3 - added ManufacturerDto and logic for it.

* PO-3 - added ProductTypeDto and logic for it.

* PO-3 - added ProductDto and logic for it.

* PO-3 - added ProductDto and logic for it.

* PO-3 -  hotfix of a bug.

* PO-3 - added testing properties to run unit tests on in-memory H2.

* PO-3 - finished ProductDto and all required staff for it.

* PO-3 - added ProductPriceDto and all required for it staff, also made some improvements in unit tests.

* PO-3 - small unit tests improvements.

* PO-3 - added WishListDto and all it required staff.

* Update ProductDtoConverterImplTest.java

* Update ProductPriceDtoConverterImplTest.java

* Update StoreDtoConverterImplTest.java

* Update UserDtoConverterImplTest.java

* Update WishListDtoConverterImplTest.java

* Update application.properties

+++ main.sql
@@ -38,13 +38,13 @@ CREATE TABLE IF NOT EXISTS `user`
     `email`      VARCHAR(70) NOT NULL,
     `birth`      DATE        NOT NULL,
     `password`   VARCHAR(50) NOT NULL,
-    `role_id`    INT         NOT NULL,
+    `role_id`    TINYINT     NOT NULL,
     PRIMARY KEY (`id`)
 );
 
 CREATE TABLE IF NOT EXISTS `user_role`
 (
-    `id`   INT         NOT NULL AUTO_INCREMENT,
+    `id`   TINYINT         NOT NULL AUTO_INCREMENT,
     `name` VARCHAR(50) NOT NULL,
     PRIMARY KEY (`id`)
 );
@@ -106,6 +106,8 @@ ALTER TABLE `product_price`
 
 ALTER TABLE `product_price`
     ADD CONSTRAINT `product_price_fk1` FOREIGN KEY (`store_id`) REFERENCES `store` (`id`);
+
+/* Prevent inserting more than 3 user roles (user, admin and god) into the user_role table. */
 
 DELIMITER $$
 CREATE TRIGGER before_insert_trigger
